+++ bigframes/core/compile/sqlglot/__init__.py
@@ -15,7 +15,6 @@ from __future__ import annotations
 
 from bigframes.core.compile.sqlglot.compiler import SQLGlotCompiler
 import bigframes.core.compile.sqlglot.expressions.array_ops  # noqa: F401
-import bigframes.core.compile.sqlglot.expressions.binary_compiler  # noqa: F401
 import bigframes.core.compile.sqlglot.expressions.blob_ops  # noqa: F401
 import bigframes.core.compile.sqlglot.expressions.comparison_ops  # noqa: F401
 import bigframes.core.compile.sqlglot.expressions.date_ops  # noqa: F401
