+++ bigframes/core/compile/sqlglot/expressions/blob_ops.py
@@ -21,6 +21,7 @@ from bigframes.core.compile.sqlglot.expressions.typed_expr import TypedExpr
 import bigframes.core.compile.sqlglot.scalar_compiler as scalar_compiler
 
 register_unary_op = scalar_compiler.scalar_op_compiler.register_unary_op
+register_binary_op = scalar_compiler.scalar_op_compiler.register_binary_op
 
 
 @register_unary_op(ops.obj_fetch_metadata_op)
@@ -31,3 +32,8 @@ def _(expr: TypedExpr) -> sge.Expression:
 @register_unary_op(ops.ObjGetAccessUrl)
 def _(expr: TypedExpr) -> sge.Expression:
     return sge.func("OBJ.GET_ACCESS_URL", expr.expr)
+
+
+@register_binary_op(ops.obj_make_ref_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    return sge.func("OBJ.MAKE_REF", left.expr, right.expr)
