+++ bigframes/core/compile/sqlglot/expressions/comparison_ops.py
@@ -19,12 +19,13 @@ import typing
 import pandas as pd
 import sqlglot.expressions as sge
 
+from bigframes import dtypes
 from bigframes import operations as ops
 from bigframes.core.compile.sqlglot.expressions.typed_expr import TypedExpr
 import bigframes.core.compile.sqlglot.scalar_compiler as scalar_compiler
-import bigframes.dtypes as dtypes
 
 register_unary_op = scalar_compiler.scalar_op_compiler.register_unary_op
+register_binary_op = scalar_compiler.scalar_op_compiler.register_binary_op
 
 
 @register_unary_op(ops.IsInOp, pass_op=True)
@@ -53,7 +54,76 @@ def _(expr: TypedExpr, op: ops.IsInOp) -> sge.Expression:
     )
 
 
+@register_binary_op(ops.eq_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    left_expr = _coerce_bool_to_int(left)
+    right_expr = _coerce_bool_to_int(right)
+    return sge.EQ(this=left_expr, expression=right_expr)
+
+
+@register_binary_op(ops.eq_null_match_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    left_expr = left.expr
+    if right.dtype != dtypes.BOOL_DTYPE:
+        left_expr = _coerce_bool_to_int(left)
+
+    right_expr = right.expr
+    if left.dtype != dtypes.BOOL_DTYPE:
+        right_expr = _coerce_bool_to_int(right)
+
+    sentinel = sge.convert("$NULL_SENTINEL$")
+    left_coalesce = sge.Coalesce(
+        this=sge.Cast(this=left_expr, to="STRING"), expressions=[sentinel]
+    )
+    right_coalesce = sge.Coalesce(
+        this=sge.Cast(this=right_expr, to="STRING"), expressions=[sentinel]
+    )
+    return sge.EQ(this=left_coalesce, expression=right_coalesce)
+
+
+@register_binary_op(ops.ge_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    left_expr = _coerce_bool_to_int(left)
+    right_expr = _coerce_bool_to_int(right)
+    return sge.GTE(this=left_expr, expression=right_expr)
+
+
+@register_binary_op(ops.gt_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    left_expr = _coerce_bool_to_int(left)
+    right_expr = _coerce_bool_to_int(right)
+    return sge.GT(this=left_expr, expression=right_expr)
+
+
+@register_binary_op(ops.lt_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    left_expr = _coerce_bool_to_int(left)
+    right_expr = _coerce_bool_to_int(right)
+    return sge.LT(this=left_expr, expression=right_expr)
+
+
+@register_binary_op(ops.le_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    left_expr = _coerce_bool_to_int(left)
+    right_expr = _coerce_bool_to_int(right)
+    return sge.LTE(this=left_expr, expression=right_expr)
+
+
+@register_binary_op(ops.ne_op)
+def _(left: TypedExpr, right: TypedExpr) -> sge.Expression:
+    left_expr = _coerce_bool_to_int(left)
+    right_expr = _coerce_bool_to_int(right)
+    return sge.NEQ(this=left_expr, expression=right_expr)
+
+
 # Helpers
 def _is_null(value) -> bool:
     # float NaN/inf should be treated as distinct from 'true' null values
     return typing.cast(bool, pd.isna(value)) and not isinstance(value, float)
+
+
+def _coerce_bool_to_int(typed_expr: TypedExpr) -> sge.Expression:
+    """Coerce boolean expression to integer."""
+    if typed_expr.dtype == dtypes.BOOL_DTYPE:
+        return sge.Cast(this=typed_expr.expr, to="INT64")
+    return typed_expr.expr
