+++ tests/unit/core/compile/sqlglot/expressions/snapshots/test_json_ops/test_json_set/out.sql
