+++ tests/unit/core/compile/sqlglot/expressions/snapshots/test_string_ops/test_add_string/out.sql
