+++ tests/unit/core/compile/sqlglot/expressions/test_blob_ops.py
@@ -29,3 +29,8 @@ def test_obj_get_access_url(scalar_types_df: bpd.DataFrame, snapshot):
     blob_s = scalar_types_df["string_col"].str.to_blob()
     sql = blob_s.blob.read_url().to_frame().sql
     snapshot.assert_match(sql, "out.sql")
+
+
+def test_obj_make_ref(scalar_types_df: bpd.DataFrame, snapshot):
+    blob_df = scalar_types_df["string_col"].str.to_blob()
+    snapshot.assert_match(blob_df.to_frame().sql, "out.sql")
